requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pandas==2.2.3
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.6
scikit-learn==1.8.0
joblib==1.5.3
pytest==9.0.3
httpx==0.28.1
